*Ajuste de Calculo de Importe para considerar descuentos; Se modifica formato FOFOE para incluir total de Descuentos

+++ assets/js/ie/SeleccionarFormaPago/DetalleCamposClinicos/index.js
@@ -6,9 +6,9 @@ import DetalleMontoPago from "../DetalleMontoPago";
 const DetalleCamposClinicos = ({ camposClinicos, solicitud }) => {
 
   function getTotalBySolicitud() {
-    let total = 0;
+    let total = 0.0;
     for(let camposClinico of camposClinicos) {
-      total += parseInt(camposClinico.montoPagar)
+      total += parseFloat(camposClinico.montoPagar)
     }
 
     return moneyFormat(total)
